Fix and enhancement
Fixed:
Encoding error when writing search data to html file.

Enhanced:
Script now checks whether the data file exists before writing a blank
file to disk. Now if the file is interrupted or stopped the user can
start from where they left off by just running the script again.

+++ scrape_bbc_recipes.py
@@ -47,10 +47,10 @@ def add_data_to_indexhtml():
 
     # Add the search data to the html file
     # The read data is sliced because the last character will be a delimiter
-    with open('temp_data.txt', 'r') as data_file:
+    with open('temp_data.txt', 'r', encoding='utf-8') as data_file:
         with open('search.html', 'w', encoding='utf-8') as html_file:
             html_file.write(html_data[:insertion_index + 20] +
-                            data_file.read()[:-1] +
+                            data_file.read()[:-3] +
                             html_data[insertion_index + 20:])
 
 
@@ -164,8 +164,9 @@ def save_pages(css_links):
         css_links[i] = os.path.join('css', link.split('/')[-1])
 
     # Create a file that will hold each page and key information
-    with open('temp_data.txt', 'w') as f:
-        f.write('')
+    if not os.path.isfile('temp_data.txt'):
+        with open('temp_data.txt', 'w') as f:
+            f.write('')
 
     # Cycle through the sitemap grabbing each recipe
     with open('bbc_sitemap.txt', 'r') as f:
